added a few new methods
