--- lib/user.rb
@@ -1,17 +1,0 @@
-class User
-  @@all = []
-
-  def self.all
-    return @@all
-  end
-
-  ####
-  def initialize(name)
-    @name = name
-    @
-    @@all << self
-  end
-
-
-
-end
